move methods down in to editor
